+++ flytestdlib/database/db.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/flyteorg/flyte/flytestdlib/logger"
 	"github.com/go-gormigrate/gormigrate/v2"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
+
+	"github.com/flyteorg/flyte/flytestdlib/logger"
 )
 
 // GetDB uses the dbConfig to create gorm DB object. If the db doesn't exist for the dbConfig then a new one is created
